Fix subreddit naming issue.
Fix an issue in which if a user entered subreddit name did not case
match the subreddit name as it is on reddit, the reddit name would be
used to create the subreddit download folder.  This caused an issue when
trying to open the folder from the downloader as the paths did not match
case correctly.