Add data path builder and change ObjectStateHandler to use it
Create a function in SystemUtil that will create and return a path to
the data directory to be used by the application based on the users os.
This directory will be used for storing save files and log information.
Also modified the ObjectStateHandler's get save file method to use the
new data directory method from SystemUtil.