Add `ensure_dict_defaults` method and unit tests
When new fields are added to default dicts in the settings manager and an existing dict is loaded from a config file, it will not have the new fields present until the user saves the settings.  This adds a method to add new default fields to those existing loaded dicts so the proper default values of newly added settings are considered before any user interaction.